Move kafka sink related error handling to planner
This helps simplify everything in this PR, at the expense of adding a little bit
of blocking logic someplace we shouldn't keep it long term
